requirements.txt
openai
streamlit
httpx
trafilatura
beautifulsoup4
pandas
pydantic
python-dotenv
tldextract
